% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr}
\alias{irr}
\title{Internal rate of return}
\usage{
irr(cf, interval = c(-100, 100))
}
\arguments{
\item{cf}{Vector of cash flows}

\item{interval}{a vector containing the end-points of the interval to be searched for the minimum.
used if you are using this function as part of an internal rate of return calculation)}
}
\description{
Internal rate of return
}
\details{
for a given stream of cash flows, return the internal rate of return
}
