% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv.R
\name{npv}
\alias{npv}
\title{Net present value}
\usage{
npv(discount, cf, return_abs = FALSE)
}
\arguments{
\item{discount}{Discount rate, as a percentage (eg "7" means 7%)}

\item{cf}{Vector of cash flows}

\item{return_abs}{Whether or not to return the absolute value of the net present value (only likely to be
used if you are using this function as part of an internal rate of return calculation)}
}
\description{
Net present value
}
\details{
for a given stream of cash flows, number of years and discount rate, return the net
present value
}
